Update SARIMAX training code for resuming

+++ songdo_arima/src/songdo_arima/sarimax_training.py
@@ -58,44 +58,56 @@ def train(config: HyperParams) -> dict:
 
 
 def train_sensor(train_data: pd.Series, valid_data: pd.Series, model_output_dir: str):
-    # auto_arima 모델 학습
-    model: ARIMA = auto_arima(
-        train_data,
-        start_p=0,
-        max_p=3,
-        start_q=0,
-        max_q=3,
-        d=None,  # 'd' 값을 자동으로 결정
-        start_P=0,
-        max_P=1,
-        start_Q=0,
-        max_Q=1,
-        D=None,  # 'D' 값을 자동으로 결정
-        m=24,  # 계절 주기 (시간별 데이터의 일일 계절성)
-        seasonal=True,
-        stepwise=True,  # 스텝와이즈 알고리즘 사용으로 계산 시간 단축
-        suppress_warnings=True,
-        error_action="ignore",
-        trace=False,  # 진행 상황을 보려면 True로 설정
-    )
-
-    # 검증 데이터에 대한 예측
-    n_periods = len(valid_data)
-    forecast = model.predict(n_periods=n_periods)
-    forecast = pd.Series(forecast, index=valid_data.index)
-
-    # MAE 및 RMSE 계산
-    mae = mean_absolute_error(valid_data, forecast)
-    rmse = root_mean_squared_error(valid_data, forecast)
-    print(f"센서 {train_data.name}의 Auto ARIMA 모델{model.get_params()} MAE: {mae}")
-
-    best_result = BestARIMA(model)
-    best_result.result["MAE"] = mae
-    best_result.result["RMSE"] = rmse
-
-    # 모델 저장
     model_output_path = os.path.join(model_output_dir, f"{train_data.name}.pkl")
-    with open(model_output_path, "wb") as pkl:
-        pickle.dump(best_result, pkl)
+
+    if not os.path.exists(model_output_path):
+        # auto_arima 모델 학습
+        model: ARIMA = auto_arima(
+            train_data,
+            start_p=0,
+            max_p=3,
+            start_q=0,
+            max_q=3,
+            d=None,  # 'd' 값을 자동으로 결정
+            start_P=0,
+            max_P=1,
+            start_Q=0,
+            max_Q=1,
+            D=None,  # 'D' 값을 자동으로 결정
+            m=24,  # 계절 주기 (시간별 데이터의 일일 계절성)
+            seasonal=True,
+            stepwise=True,  # 스텝와이즈 알고리즘 사용으로 계산 시간 단축
+            suppress_warnings=True,
+            error_action="ignore",
+            trace=False,  # 진행 상황을 보려면 True로 설정
+        )
+
+        # 검증 데이터에 대한 예측
+        n_periods = len(valid_data)
+        forecast = model.predict(n_periods=n_periods)
+        forecast = pd.Series(forecast, index=valid_data.index)
+
+        # MAE 및 RMSE 계산
+        mae = mean_absolute_error(valid_data, forecast)
+        rmse = root_mean_squared_error(valid_data, forecast)
+
+        best_result = BestARIMA(model)
+        best_result.result["MAE"] = mae
+        best_result.result["RMSE"] = rmse
+
+        # 모델 저장
+        with open(model_output_path, "wb") as pkl:
+            pickle.dump(best_result, pkl)
+            order = model.order
+            seasonal_order = model.seasonal_order
+            print(f"센서 {train_data.name} is trained: {order} {seasonal_order}")
+    else:
+        with open(model_output_path, "rb") as pkl:
+            best_result: BestARIMA = pickle.load(pkl)
+            mae = best_result.result["MAE"]
+            rmse = best_result.result["RMSE"]
+            order = best_result.best_model.order
+            seasonal_order = best_result.best_model.seasonal_order
+            print(f"센서 {train_data.name} is already trained: {order} {seasonal_order}")
 
     return mae, rmse
